fix[aula 3]: ex02

+++ python_para_dados/aula 3/ex02.py
@@ -1,4 +1,4 @@
-percentual = int(input("Digite o percentual de crescimento: "))
+percentual = float(input("Digite o percentual de crescimento: "))
 
 if(percentual >= 0):
     print("Houve crescimento")
